separated secrets to secrets.js
